+++ packages/react-html/src/server/components/tests/Serialize.test.tsx
@@ -30,7 +30,7 @@ describe('<Serialize />', () => {
 
       expect(serialize).toContainReactComponent('script', {
         dangerouslySetInnerHTML: {
-          __html: serializeJavaScript(data, {isJSON: true}),
+          __html: serializeJavaScript({data}, {isJSON: true}),
         },
       });
     });