+++ packages/react-html/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+- Fixes an issue where passing a string to Serialize component would break serialization as it fails to properly JSON.parse
 
 ## [9.3.0] - 2020-03-13